+++ test/specs/openapi/conftest.py
@@ -1,10 +1,9 @@
 from copy import deepcopy
 
-import jsonschema
 import pytest
 
 import schemathesis
-from schemathesis.specs.openapi.definitions import OPENAPI_30, SWAGGER_20
+from schemathesis.specs.openapi.definitions import OPENAPI_30_VALIDATOR, SWAGGER_20_VALIDATOR
 
 
 def make_object_schema(is_loose=False, **properties):
@@ -93,7 +92,7 @@ def make_openapi_2_schema(empty_open_api_2_schema):
                 "responses": {"200": {"description": "OK"}},
             }
         }
-        jsonschema.validate(schema, SWAGGER_20)
+        SWAGGER_20_VALIDATOR.validate(schema)
         return schema
 
     return maker
@@ -127,7 +126,7 @@ def make_openapi_3_schema(empty_open_api_3_schema):
         if parameters is not None:
             definition["parameters"] = parameters
         schema["paths"]["/users"] = {"post": definition}
-        jsonschema.validate(schema, OPENAPI_30)
+        OPENAPI_30_VALIDATOR.validate(schema)
         return schema
 
     return maker